v3/v3_app/requirements.txt
# Core dependencies
transformers>=4.51.3
torch>=2.3.0
numpy>=2.2.2
scipy>=1.13.0
streamlit>=1.36.0

# LangChain ecosystem
langchain>=0.2.11
langgraph>=0.2.55
langchain-core>=0.2.4
langchain-community>=0.3.14
langchain-openai==0.2.14
langchain-anthropic>=0.3.8
langchain-qdrant>=0.2.0
langgraph==0.2.61
langchain-huggingface>=0.0.5
langchain-cohere>=0.1.0

tavily-python>=0.5.0
xmltodict>=0.14.2

# Retrieval and ranking
rank_bm25>=0.2.2

# Hugging Face
huggingface-hub>=0.23.2

# Utilities
python-dotenv>=1.0.1
arxiv>=2.1.3
tiktoken>=0.7.0
unstructured>=0.14.8
ipython>=7.0.0

